LPD-43574 - Refactor. Move ckeditor4 styles under its own folder
