+++ modules/apps/frontend-editor/frontend-editor-ckeditor-web/src/main/resources/META-INF/resources/js/ckeditor4/BaseEditor.js
@@ -16,7 +16,7 @@ import React, {
 	useState,
 } from 'react';
 
-import '../../css/main.scss';
+import '../../css/ckeditor4/main.scss';
 
 const BASEPATH = '/o/frontend-editor-ckeditor-web/ckeditor/';
 const CONTEXT_URL = Liferay.ThemeDisplay.getPathContext();
